pyproject.toml
[project]
name = "mat-dnf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "giprayogo", email = "prayogo.genki.imam@advancesoft.jp" }]
requires-python = ">=3.12"
dependencies = [
    "cupy-cuda12x>=13.3.0",
    "equinox>=0.11.11",
    "jax[cuda12]>=0.5.0",
    "jaxtyping>=0.2.37",
    "numpy>=2.2.1",
    "optax>=0.2.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["pyright>=1.1.391", "pytest>=8.3.4", "pytest-cov>=6.0.0", "ruff>=0.9.1"]
notebook = ["ipykernel>=6.29.5"]
benchmark = ["py-spy>=0.4.0"]
docs = [
    "pydata-sphinx-theme>=0.16.1",
    "sphinx>=8.2.0",
    "sphinx-autoapi>=3.6.0",
]

[tool.pyright]
typeCheckingMode = "strict"
reportMissingTypeStubs = "none"
reportUnknownArgumentType = "none"
reportUnknownLambdaType = "none"
reportUnknownMemberType = "none"
reportUnknownParameterType = "none"
reportUnknownVariableType = "none"

[tool.ruff.lint]
select = ["D", "A", "ICN", "PL", "PERF"]
ignore = ["PLR0913"]

[tool.ruff.lint.per-file-ignores]
"tests/conftest.py" = ["D103"]

[tool.ruff.lint.pydocstyle]
convention = "google"
